delete agreements and skzi

+++ server/src/agreements/agreements.service.ts
@@ -3,6 +3,8 @@ import { agreementsTransform } from './agreements.transform'
 import { CreateAgreementDto } from './dto/create-agreement.dto'
 import { UpdateAgreementDto } from './dto/update-agreement.dto'
 import { PaginationDto } from '../dto/pagination.dto'
+import { skziUnitsService } from '../skzi-units/skzi-units.service'
+import { actsService } from '../acts/acts.service'
 import { ApiError } from '../exceptions/api-error'
 
 class AgreementsService {
@@ -41,10 +43,27 @@ class AgreementsService {
   }
 
   async remove(id: number) {
+    const skziUnitsCount = await skziUnitsService.count({ agreementId: id })
+    const actsCount = await actsService.count({ agreementId: id })
+
+    if(skziUnitsCount || actsCount) {
+      throw ApiError.BadRequest('Невозможно удалить соглашение, т.к. существуют связанные СКЗИ или акты')
+    }
+
+    const agreement = await agreementsRepository.getById(id)
+    if(!agreement) {
+      throw ApiError.NotFound('Соглашение не найдено')
+    }
+
+    if(agreement.parentId) {
+      throw ApiError.BadRequest('Невозможно удалить соглашение, т.к. существует родительское соглашение')
+    }
+
     const isDeleted = await agreementsRepository.remove(id)
     if(!isDeleted) {
       throw ApiError.NotFound('Соглашение не найдено')
     }
+
     return isDeleted
   }
 }
